terraform-vernemq-iot-core/requirements.txt
paho-mqtt==1.6.1
pytz==2021.3
